+++ Entangle/entangle-android/res/layout/super_profile.xml
@@ -6,7 +6,7 @@
     android:fillViewport="true" >
 
     <LinearLayout
-        android:id="@+id/profile"
+        android:id="@+id/profile_layout"
         android:layout_width="fill_parent"
         android:layout_height="fill_parent"
         android:layout_alignParentTop="true"
